+++ client/src/component styles/welcome.css
