+++ client/src/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import App from './app';
+
+
+ReactDOM.render(<App />, document.getElementById('root'));
